refactor: Move activity components in a dedicated folder
