+++ app/components/activities/list_component.html.erb
@@ -1,6 +1,6 @@
 <ul class="op-activity-list">
   <% @events.each do |event| -%>
     <% journal = @journals_by_id[event.event_id] -%>
-    <%= render(ActivityItemComponent.new(event:, journal:, display_user: @display_user)) %>
+    <%= render(Activities::ItemComponent.new(event:, journal:, display_user: @display_user)) %>
   <% end -%>
 </ul>
