+++ app/components/activities/list_component.rb
@@ -28,33 +28,11 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-require "rails_helper"
-
-RSpec.describe ActivityItemComponent, type: :component do
-  let(:event) do
-    Activities::Event.new(
-      event_title: "Event Title",
-      event_description: "something",
-      event_datetime: journal.created_at,
-      project_id: project.id,
-      project:,
-      event_path: "/project/123"
-    )
-  end
-  let(:project) { build_stubbed(:project) }
-  let(:journal) { build_stubbed(:work_package_journal) }
-
-  it 'renders the title escaped' do
-    event.event_title = 'Hello <b>World</b>!'
-    render_inline(described_class.new(event:, journal:))
-
-    expect(page).to have_css('.op-activity-list--item-title', text: 'Hello <b>World</b>!')
-  end
-
-  it 'renders the project name to which the event belongs, escaped' do
-    event.project.name = 'Project <b>name</b> with HTML'
-    render_inline(described_class.new(event:, journal:))
-
-    expect(page).to have_css('.op-activity-list--item-title', text: '(Project: Project <b>name</b> with HTML)')
+class Activities::ListComponent < ViewComponent::Base
+  def initialize(events:, journals_by_id:, display_user: true)
+    super()
+    @events = events.sort { |x, y| y.event_datetime <=> x.event_datetime }
+    @journals_by_id = journals_by_id
+    @display_user = display_user
   end
 end
