+++ app/views/activities/index.html.erb
@@ -37,7 +37,7 @@ See COPYRIGHT and LICENSE files for more details.
   <% @events_by_day.sort_by { |day, _events| day }.reverse_each do |day, events| %>
     <h3 class="op-activity-days--header"><%= format_activity_day(day) %></h3>
 
-    <%= render(ActivityListComponent.new(events:, journals_by_id: @journals_by_id)) %>
+    <%= render(Activities::ListComponent.new(events:, journals_by_id: @journals_by_id)) %>
   <% end -%>
 </div>
